chore: moving the address service and renaming it references
